x11: wait until the window is mapped
This blocks everything, until the window is actually reported as mapped.

This fixes the race condition between VO initialization and mapping the
window, which resulted in possibly different window sizes, leading to an
immediate redraw, visible as flashing.

This does not fix the (probably unfixable) race condition between
changing the window state and redrawing in the right moment with the
right size. You'll need Wayland to get this.